first pass: move Macro to dbt/artifacts
