+++ core/dbt/artifacts/resources/__init__.py
@@ -2,3 +2,5 @@ from dbt.artifacts.resources.base import BaseArtifactNode
 
 # alias to latest resource definitions
 from dbt.artifacts.resources.v1.documentation import Documentation
+from dbt.artifacts.resources.v1.macro import Macro, MacroDependsOn, MacroArgument
+from dbt.artifacts.resources.v1.docs import Docs
